M    Makefile.in M    src/buildnumber.h M    src/path_resolution.c M    src/tagsistant.c M    src/fuse_operations/operations.h M    src/fuse_operations/write.c M    src/fuse_operations/open.c M    src/fuse_operations/rename.c M    src/fuse_operations/readdir.c M    src/fuse_operations/getattr.c M    src/fuse_operations/read.c M    src/path_resolution.h M    src/tagsistant.h M    src/debug.h M    src/sql.c M    src/plugins/Makefile.in M    src/utils.c M    src/Makefile.in M    .cproject M    autom4te.cache/requests M    autom4te.cache/output.0 M    autom4te.cache/traces.0 M    .autotools M    aclocal.m4 M    .settings/org.eclipse.cdt.codan.core.prefs M    configure

+++ src/fuse_operations/getattr.c
@@ -93,7 +93,7 @@ int tagsistant_getattr(const char *path, struct stat *stbuf)
 	tagsistant_errno = errno;
 
 	// delete non existent objects
-	if (qtree->is_taggable && res == -1 && tagsistant_errno == ENOENT) {
+	if (0 && qtree->is_taggable && res == -1 && tagsistant_errno == ENOENT) {
 		tagsistant_query("delete from objects where inode = %d", qtree->conn, NULL, NULL, qtree->inode);
 		tagsistant_query("delete from tagging where inode = %d", qtree->conn, NULL, NULL, qtree->inode);
 	}
